Changes for OBF-226 Create tool to compare archetypes
This requires an upgrade to castor 1.3.2

git-svn-id: svn://localhost/openvpms/etl/trunk@5542 9e4fa352-aafb-0310-9071-df624684024b

+++ etl/src/java/org/openvpms/etl/kettle/LoaderPluginDialog.java
@@ -373,7 +373,7 @@ public class LoaderPluginDialog extends BaseStepDialog
             idName.setText(mappings.getIdColumn());
         }
 
-        batchSize.setText(Integer.toString(mappings.getBatchSize()));
+        batchSize.setText(Long.toString(mappings.getBatchSize()));
         generateLookups.setSelection(mappings.getGenerateLookups());
 
         skipProcessed.setSelection(mappings.getSkipProcessed());
